+++ app/models/users/struct.go
@@ -3,35 +3,10 @@ package users
 import (
 	"database/sql"
 	"time"
-
-	"github.com/daison12006013/gorvel/pkg/paginate"
 )
 
 const Table = "users"
 const PrimaryKey = "id"
-
-type Input struct {
-	Visible       bool
-	Placeholder   interface{}
-	Value         interface{}
-	CanSearch     bool
-	SearchColumn  []string
-	SearchPattern string
-}
-
-type Header struct {
-	Name  interface{}
-	Input Input
-}
-
-type SearchableTable struct {
-	Paginate paginate.Paginate
-
-	Headers     []Header
-	Params      map[string]string
-	OrderByCol  *string
-	OrderBySort *string
-}
 
 type Attributes struct {
 	ID              uint           `gorm:"primarykey" json:"id"`
@@ -43,31 +18,3 @@ type Attributes struct {
 	CreatedAt       time.Time      `gorm:"column:created_at" json:"created_at"`
 	UpdatedAt       time.Time      `gorm:"column:updated_at" json:"updated_at"`
 }
-
-// func (t *Attributes) GetCreatedAt() time.Time {
-// 	helpers.DD(t.CreatedAt)
-// 	tp, err := time.Parse("2006-01-02 3:04PM", t.CreatedAt)
-// 	if err != nil {
-// 		logger.Fatal(err)
-// 		panic(err)
-// 	}
-// 	return tp
-// }
-
-// func (t *Attributes) GetUpdatedAt() time.Time {
-// 	tp, err := time.Parse("2006-01-02 3:04PM", t.CreatedAt)
-// 	if err != nil {
-// 		logger.Fatal(err)
-// 		panic(err)
-// 	}
-// 	return tp
-// }
-
-// func (t *Attributes) GetEmailVerifiedAt() time.Time {
-// 	tp, err := time.Parse("2006-01-02 3:04PM", t.CreatedAt)
-// 	if err != nil {
-// 		logger.Fatal(err)
-// 		panic(err)
-// 	}
-// 	return tp
-// }
