+++ resources/views/users/_table.go.html
@@ -1,82 +1,77 @@
 {{ define "component" }}
 <div class="flex flex-col">
-  <div class="flex justify-between mb-2">
-    <div class="text-xs inline-block my-auto">
+    <div class="flex justify-between mb-2">
+      <div class="text-xs inline-block my-auto">
         <span class="border-dotted border-gray-400 border-b"><a href="/">Home</a></span>
         <i class="mx-2">&#8725;</i>
         <span>{{ $.title }}</span>
-    </div>
+      </div>
 
-    <form method="GET" class="w-1/4">
-      {{ range $idx, $header := $.data.Headers }}
-        {{ if eq $header.Name "search" }}
-          <input name="{{ $header.Name }}" value="{{ $header.Input.Value }}" type="text" placeholder="{{ $header.Input.Placeholder }}" class="sm:rounded-lg md:rounded-md outline-0 px-3 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider w-full" />
-        {{ else }}
-          <input type="hidden" name="{{ $header.Name }}" value="{{ $header.Input.Value }}" />
-        {{ end }}
-      {{ end }}
-    </form>
-  </div>
-  <div class="-my-2 overflow-x-auto sm:-mx-6 lg:-mx-8">
-    <div class="py-2 align-middle inline-block min-w-full sm:px-6 lg:px-8">
-      <div class="shadow overflow-hidden border-b border-gray-200 sm:rounded-lg">
-        <table class="min-w-full divide-y divide-gray-200">
-          <thead class="bg-gray-50">
-            <tr>
-              {{ range $idx, $header := $.data.Headers }}
-                {{ if $header.Input.Visible }}
-                  <th scope="col" class="text-left text-xs font-medium text-gray-500 tracking-wider">
-                    <form method="GET">
-                      <input
-                        class="bg-transparent outline-0 px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider"
-                        name="search[{{ $header.Name }}]"
-                        value="{{ $header.Input.Value }}"
-                        placeholder="{{ $header.Input.Placeholder }}"
-                      />
-
-                      {{ range $idx2, $header2 := $.data.Headers }}
-                        {{ if not (eq $header.Name $header2.Name) }}
-                          <input type="hidden" name="{{ $header2.Name }}" value="{{ $header2.Input.Value }}" />
-                        {{ end }}
-                      {{ end }}
-                    </form>
-                  </th>
-                {{ end }}
-              {{ end }}
-              <th scope="col" class="text-center text-xs font-medium text-gray-500 uppercase tracking-wider">Action</th>
-            </tr>
-          </thead>
-          <tbody class="bg-white divide-y divide-gray-200">
-            {{ range $idx, $record := $.data.Paginate.Items }}
-            <tr>
-              <td class="px-6 py-4 whitespace-nowrap">
-                <div class="flex items-center">
-                  <div class="flex-shrink-0 h-10 w-10">
-                    <img class="h-10 w-10 rounded-full" src="https://abs.twimg.com/sticky/default_profile_images/default_profile_normal.png" alt="">
-                  </div>
-                  <div class="ml-4">
-                    <div class="text-sm font-medium text-gray-900">{{ $record.Name }}</div>
-                  </div>
-                </div>
-              </td>
-              <td class="px-6 py-4 whitespace-nowrap">
-                <div class="text-sm text-gray-500">{{ $record.Email }}</div>
-              </td>
-              <td class="px-6 py-4 whitespace-nowrap text-center text-sm font-medium">
-                <a href="/users/{{ $record.ID }}" class="text-indigo-600 hover:text-indigo-900 px-2">View</a>
-                <form class="px-2 inline-block" action="/users/{{ $record.ID }}/delete" method="POST">
-                  {{ $.csrfField }}
-                  <button class="text-red-600 hover:text-red-900">Delete</button>
-                </form>
-              </td>
-            </tr>
+      <form method="GET" class="w-1/4">
+        {{ range $idx, $header := $.data.Headers }}
+            {{ if eq $header.Name "search" }}
+                <input name="{{ $header.Name }}" value="{{ $header.Input.Value }}" type="text" placeholder="{{ $header.Input.Placeholder }}" class="sm:rounded-lg md:rounded-md outline-0 px-3 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider w-full" />
+            {{ else }}
+                <input type="hidden" name="{{ $header.Name }}" value="{{ $header.Input.Value }}" />
             {{ end }}
+        {{ end }}
+      </form>
+    </div>
+    <div class="-my-2 overflow-x-auto sm:-mx-6 lg:-mx-8">
+      <div class="py-2 align-middle inline-block min-w-full sm:px-6 lg:px-8">
+        <div class="shadow overflow-hidden border-b border-gray-200 sm:rounded-lg">
+          <table class="min-w-full divide-y divide-gray-200">
+            <thead class="bg-gray-50">
+              <tr>
+                {{ range $idx, $header := $.data.Headers }}
+                    {{ if $header.Input.Visible }}
+                    <th scope="col" class="text-left text-xs font-medium text-gray-500 tracking-wider">
+                    <form method="GET">
+                        <input class="bg-transparent outline-0 px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider" name="search[{{ $header.Name }}]" value="{{ $header.Input.Value }}" placeholder="{{ $header.Input.Placeholder }}" />
 
-            <!-- More people... -->
-          </tbody>
-        </table>
+                        {{ range $idx2, $header2 := $.data.Headers }}
+                            {{ if not (eq $header.Name $header2.Name) }}
+                                <input type="hidden" name="{{ $header2.Name }}" value="{{ $header2.Input.Value }}" />
+                            {{ end }}
+                        {{ end }}
+                    </form>
+                    </th>
+                    {{ end }}
+                {{ end }}
+                <th scope="col" class="text-center text-xs font-medium text-gray-500 uppercase tracking-wider">Action</th>
+              </tr>
+            </thead>
+            <tbody class="bg-white divide-y divide-gray-200">
+              {{ range $idx, $record := $.data.Paginate.Items }}
+              <tr>
+                <td class="px-6 py-4 whitespace-nowrap">
+                  <div class="flex items-center">
+                    <div class="flex-shrink-0 h-10 w-10">
+                      <img class="h-10 w-10 rounded-full" src="https://abs.twimg.com/sticky/default_profile_images/default_profile_normal.png" alt="">
+                    </div>
+                    <div class="ml-4">
+                      <div class="text-sm font-medium text-gray-900">{{ $record.Name }}</div>
+                    </div>
+                  </div>
+                </td>
+                <td class="px-6 py-4 whitespace-nowrap">
+                  <div class="text-sm text-gray-500">{{ $record.Email }}</div>
+                </td>
+                <td class="px-6 py-4 whitespace-nowrap text-center text-sm font-medium">
+                  <a href="/users/{{ $record.ID }}" class="text-indigo-600 hover:text-indigo-900 px-2">View</a>
+                  <form class="px-2 inline-block" action="/users/{{ $record.ID }}/delete" method="POST">
+                    {{ $.csrfField }}
+                    <button class="text-red-600 hover:text-red-900">Delete</button>
+                  </form>
+                </td>
+              </tr>
+              {{ end }}
+
+              <!-- More people... -->
+            </tbody>
+          </table>
+        </div>
       </div>
     </div>
-  </div>
 </div>
 {{ end }}
