+++ resources/views/users/show.go.html
@@ -1,11 +1,83 @@
 {{define "head"}}{{end}}
 
 {{define "body"}}
-    <body class="bg-violet-800 flex h-screen">
-        <div class="text-white m-auto">
-            <h1 class="text-8xl mb-5">
-                Show Form
-            </h1>
+<body class="bg-gray-800 flex h-screen mt-auto">
+  <div class="text-white m-auto">
+      <div class="flex justify-between mb-3">
+        <div class="text-xs inline-block my-auto">
+          <span class="border-dotted border-gray-400 border-b"><a href="/">Home</a></span>
+          <i class="mx-2">&#8725;</i>
+          <span><a href="{{ $.previousUrl }}">Users List</a></span>
+          <i class="mx-2">&#8725;</i>
+          <spa>{{ $.title }}</span>
         </div>
-    </body>
+      </div>
+      <div class="-my-2 overflow-x-auto sm:-mx-6 lg:-mx-8">
+        <div class="py-2 align-middle inline-block max-w-full sm:px-6 lg:px-8">
+          <div
+            class="p-5 bg-gray-100 shadow overflow-hidden border-b border-gray-200 sm:rounded-lg"
+          >
+            <form class="w-full">
+              <div class="flex flex-wrap -mx-3 mb-6">
+                <div class="w-1/2 md:w-full px-3 mb-6 md:mb-0">
+                  <label
+                    class="block uppercase tracking-wide text-gray-700 text-xs font-bold mb-2"
+                    for="grid-name"
+                  >
+                    Name
+                  </label>
+                  <input
+                    class="appearance-none block w-full bg-gray-200 text-gray-700 border border-red-500 rounded py-3 px-4 mb-3 leading-tight focus:outline-none focus:bg-white"
+                    id="grid-name"
+                    type="text"
+                    placeholder="John"
+                  />
+                  {{/* <p class="text-red-500 text-xs italic">
+                    Please fill out this field.
+                  </p> */}}
+                </div>
+                <div class="w-1/2 md:w-full px-3 mb-6 md:mb-0">
+                  <label
+                    class="block uppercase tracking-wide text-gray-700 text-xs font-bold mb-2"
+                    for="grid-email"
+                  >
+                    Email
+                  </label>
+                  <input
+                    class="appearance-none block w-full bg-gray-200 text-gray-700 border border-red-500 rounded py-3 px-4 mb-3 leading-tight focus:outline-none focus:bg-white"
+                    id="grid-email"
+                    type="text"
+                    placeholder="john@doe.com"
+                  />
+                  {{/* <p class="text-red-500 text-xs italic">
+                    Please fill out this field.
+                  </p> */}}
+                </div>
+              </div>
+              <div class="flex flex-wrap -mx-3 mb-6">
+                <div class="w-full px-3">
+                  <label
+                    class="block uppercase tracking-wide text-gray-700 text-xs font-bold mb-2"
+                    for="grid-password"
+                  >
+                    Password
+                  </label>
+                  <input
+                    class="appearance-none block w-full bg-gray-200 text-gray-700 border border-gray-200 rounded py-3 px-4 mb-3 leading-tight focus:outline-none focus:bg-white focus:border-gray-500"
+                    id="grid-password"
+                    type="password"
+                    placeholder="******************"
+                  />
+                  {{/* <p class="text-gray-600 text-xs italic">
+                    Make it as long and as crazy as you'd like
+                  </p> */}}
+                </div>
+              </div>
+            </form>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
+</body>
 {{end}}
